financial data fixes

+++ www/site-online/objects/Controller/OrderController.php
@@ -47,7 +47,7 @@
 
 		function readJson()
 		{
-			$query = mysql_query("SELECT id,deleted FROM local_stores WHERE id != 0",$this->connection);
+			$query = mysql_query("SELECT `id`,`deleted` FROM `local_stores` WHERE `id` != 0 AND `deleted` = 0",$this->connection);
 			date_default_timezone_set("Asia/Singapore");
 			$date = date('Y-m-d');
 			while($row = mysql_fetch_array($query))
@@ -57,18 +57,34 @@
 					$shop_id = $row['id'];
 					$file = 'receive/'.$row['id'].'.json';
 					$json = json_decode(file_get_contents($file), true);
-					foreach($json['products'] as $data)
-					{
-						$barcode = $data['barcode'];
-						$quantity = $this->decrypt($data['quantity']);
-						$sales = $this->decrypt($data['sales']);
-						$writeoff = $this->decrypt($data['write-off']);
-						if($quantity>0) {
-							mysql_query("INSERT INTO product_order(barcode, date, store_id, quantity, processed) VALUES ('$barcode', '$date', '$shop_id', '$quantity', b'0')",$this->connection);
+					if ($json != null){
+						foreach($json['products'] as $data)
+						{
+							$barcode = $data['barcode'];
+							$quantity = $this->decrypt($data['quantity']);
+							$sales = $this->decrypt($data['sales']);
+							$writeoff = $this->decrypt($data['write-off']);
+							if($quantity>0) {
+								mysql_query("INSERT INTO product_order(barcode, date, store_id, quantity, processed) VALUES ('$barcode', '$date', '$shop_id', '$quantity', b'0')",$this->connection);
+							}
+							mysql_query("INSERT INTO product_sales(barcode, date, store_id, sales, writeoff) VALUES ('$barcode', '$date', '$shop_id', '$sales', '$writeoff')",$this->connection);
 						}
-						mysql_query("INSERT INTO product_sales(barcode, date, store_id, sales, writeoff) VALUES ('$barcode', '$date', '$shop_id', '$sales', '$writeoff')",$this->connection);
+						$revenue = $this->decrypt($json['total']);
+						$sql= "SELECT * FROM `balance_sheet` WHERE `date` = CURDATE() AND `account` = 601 AND `store_id` = ".$shop_id;
+						$res = mysql_query($sql,$this->connection);
+						if (!$res) throw new Exception("Database access failed: " . mysql_error());
+						$rows = mysql_num_rows($res);
+						if  ($rows == 0) {
+							$sql = "INSERT INTO `balance_sheet` ( `date` , `account` , `store_id` , `amount` ) VALUES ( CURDATE() , 601, ".$shop_id." , ".$revenue." )";
+							$res = mysql_query($sql,$this->connection);
+							if (!$res) throw new Exception("Database access failed: " . mysql_error());
+						} else {
+							$sql = "UPDATE `balance_sheet` SET `amount` = `amount` + ".$revenue." WHERE `store_id` = ".$shop_id." AND `account` = 601 AND `date` = CURDATE()";
+							$res = mysql_query($sql,$this->connection);
+							if (!$res) throw new Exception("Database access failed: " . mysql_error());
+						}
+						unlink('receive/'.$shop_id.'.json');
 					}
-					unlink('receive/'.$shop_id.'.json');
 				}
 			}
 		}
@@ -141,11 +157,11 @@
 				$rows = mysql_num_rows($res);
 				$productCost = array();
 				for ($i=0; $i<$rows;$i++) {
-						$productCost[mysql_result($res,$j,'barcode')] = mysql_result($res,$j,'cost');
+						$productCost[mysql_result($res,$i,'barcode')] = mysql_result($res,$i,'cost');
 				}
 				$totalCost = 0;
 				for ($i=0; $i < count($barcodeShipped); $i++) {
-						$totalCost+=$productCost[$barcodeShipped[$i]['barcode']]*$barcodeShipped[$i]['quantity'];
+						$totalCost+=$productCost[$barcodeShipped[$i]['barcode']]* $this->decrypt($barcodeShipped[$i]['quantity']);
 				}
 				$sql= "SELECT * FROM `balance_sheet` WHERE `date` = CURDATE() AND `account` = 701 AND `store_id` = ".$store;
 				$res = mysql_query($sql,$this->connection);
@@ -165,14 +181,23 @@
 				$res = mysql_query($sql,$this->connection);
 				if (!$res) throw new Exception("Database access failed: " . mysql_error());
 				$rows = mysql_num_rows($res);
-				$profit = 0;
+				$revenue = 0;
 				if  ($rows == 0) {
-					$profit = 0 - $totalCost;
-				} else {
+					$revenue = 0;
+				} else {					
 					$revenue = mysql_result($res,0,'amount');
-					$profit = $revenue - $totalCost;
 				}
-				
+				$sql= "SELECT * FROM `balance_sheet` WHERE `date` = CURDATE() AND `account` = 701 AND `store_id` = ".$store;
+				$res = mysql_query($sql,$this->connection);
+				if (!$res) throw new Exception("Database access failed: " . mysql_error());
+				$rows = mysql_num_rows($res);
+				$totalCost = 0;
+				if  ($rows == 0) {
+					$totalCost = 0;
+				} else {					
+					$totalCost = mysql_result($res,0,'amount');
+				}
+				$profit = $revenue - $totalCost;
 				$sql= "SELECT * FROM `balance_sheet` WHERE `date` = CURDATE() AND `account` = 801 AND `store_id` = ".$store;
 				$res = mysql_query($sql,$this->connection);
 				if (!$res) throw new Exception("Database access failed: " . mysql_error());
@@ -182,7 +207,7 @@
 					$res = mysql_query($sql,$this->connection);
 					if (!$res) throw new Exception("Database access failed: " . mysql_error());
 				} else {
-					$sql = "UPDATE `balance_sheet` SET `amount` = `amount` + ".$profit." WHERE `store_id` = ".$store." AND `account` = 801 AND `date` = CURDATE()";
+					$sql = "UPDATE `balance_sheet` SET `amount` = ".$profit." WHERE `store_id` = ".$store." AND `account` = 801 AND `date` = CURDATE()";
 					$res = mysql_query($sql,$this->connection);
 					if (!$res) throw new Exception("Database access failed: " . mysql_error());
 				}
